Optimize GetRouteOverviewPresenter graph generation
Previously we would store the connecting lines of the data points in
a list and then run all the graphics transformations on both the data
points and the connecting lines. Now we only transform the datapoints
and derive the connecting lines from those data points. This should
provide a mild speed of up the graph generation.